Create PlatformConfiguration
