+++ src/Infrastructure/Data/ApplicationDbContext.cs
@@ -15,58 +15,13 @@ namespace Infrastructure.Data
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
-            modelBuilder.Entity<Platform>()
-                .Property(p => p.Id)
-                .ValueGeneratedOnAdd();
-
-            modelBuilder.Entity<Platform>().HasData(CreatePlatformDataSeed());
+            /* esta instrucción captura las configuraciones individuales de cada entidad que se encuentran
+             * en este mismo ensamblado. Están ubicadas en:
+             * Infrastructure/Data/Configurations/
+             */
+            modelBuilder.ApplyConfigurationsFromAssembly(typeof(ApplicationDbContext).Assembly);
 
             base.OnModelCreating(modelBuilder);
-        }
-
-        private Platform[] CreatePlatformDataSeed()
-        {
-            Platform[] newPlatforms = [
-
-                new Platform
-                {
-                    Id = 1,
-                    Name= "None",
-                    Description = "Sin plataforma"
-                },
-                new Platform
-                {
-                    Id = 2,
-                    Name= "PC",
-                    Description = "Personal Computer"
-                },
-                new Platform
-                {
-                    Id = 3,
-                    Name= "PS1",
-                    Description = "Play Station 1"
-                },
-                new Platform
-                {
-                    Id = 4,
-                    Name= "PS2",
-                    Description = "Play Station 2"
-                },
-                new Platform
-                {
-                    Id = 5,
-                    Name= "NES",
-                    Description = "Nintendo Entertainment System"
-                },
-                new Platform
-                {
-                    Id = 6,
-                    Name= "SNES",
-                    Description = "Super Nintendo Entertainment System"
-                },
-            ];
-
-            return newPlatforms;
         }
     }
 }
